requirements.txt
omegaconf
hydra-core
neptune-client
psutil
rouge-score
sacrebleu
wandb
streamlit
transformers
tokenizers
datasets
huggingface-hub
pytorch-lightning
sentencepiece